# web/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SupportNova API",
    "description": "Complaint intelligence: GenAI analysis + Python ground-truth validation.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/contact": {
      "post": {
        "tags": [
          "contact"
        ],
        "summary": "Contact",
        "description": "Order problems (sign-in required) become complaints; other topics become enquiries.",
        "operationId": "contact",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ContactIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContactOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/enquiries": {
      "get": {
        "tags": [
          "contact"
        ],
        "summary": "List Enquiries",
        "operationId": "list_enquiries",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/EnquiryStatus"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "topic",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "order_problem",
                    "product_question",
                    "business",
                    "feedback",
                    "other"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Topic"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EnquiryOut"
                  },
                  "title": "Response List Enquiries"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/enquiries/{ref}": {
      "patch": {
        "tags": [
          "contact"
        ],
        "summary": "Update Enquiry",
        "operationId": "update_enquiry",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "ref",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Ref"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EnquiryUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EnquiryOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/enquiries/{ref}/convert": {
      "post": {
        "tags": [
          "contact"
        ],
        "summary": "Convert Enquiry",
        "description": "Create a complaint from an enquiry sent by a signed-in customer (once).",
        "operationId": "convert_enquiry",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "ref",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Ref"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EnquiryOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/newsletter": {
      "post": {
        "tags": [
          "contact"
        ],
        "summary": "Newsletter",
        "description": "Demo newsletter sign-up: stored once, nothing is ever sent.",
        "operationId": "newsletter",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NewsletterIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ContactIn": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Name"
          },
          "email": {
            "type": "string",
            "maxLength": 320,
            "pattern": "^[^@\\s]+@[^@\\s]+\\.[^@\\s]+$",
            "title": "Email"
          },
          "topic": {
            "type": "string",
            "enum": [
              "order_problem",
              "product_question",
              "business",
              "feedback",
              "other"
            ],
            "title": "Topic"
          },
          "message": {
            "type": "string",
            "maxLength": 5000,
            "minLength": 10,
            "title": "Message"
          },
          "order_ref": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 20
              },
              {
                "type": "null"
              }
            ],
            "title": "Order Ref"
          },
          "website": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Website",
            "description": "Honeypot: leave empty"
          }
        },
        "type": "object",
        "required": [
          "name",
          "email",
          "topic",
          "message"
        ],
        "title": "ContactIn"
      },
      "ContactOut": {
        "properties": {
          "kind": {
            "type": "string",
            "enum": [
              "complaint",
              "enquiry",
              "ignored"
            ],
            "title": "Kind"
          },
          "reference": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reference"
          }
        },
        "type": "object",
        "required": [
          "kind",
          "reference"
        ],
        "title": "ContactOut"
      },
...
      "EnquiryOut": {
        "properties": {
          "ref": {
            "type": "string",
            "title": "Ref"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "email": {
            "type": "string",
            "title": "Email"
          },
          "topic": {
            "type": "string",
            "title": "Topic"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "status": {
            "type": "string",
            "enum": [
              "new",
              "handled"
            ],
            "title": "Status"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "handled_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Handled At"
          },
          "complaint_ref": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Complaint Ref"
          },
          "can_convert": {
            "type": "boolean",
            "title": "Can Convert"
          }
        },
        "type": "object",
        "required": [
          "ref",
          "name",
          "email",
          "topic",
          "message",
          "status",
          "created_at",
          "handled_at",
          "complaint_ref",
          "can_convert"
        ],
        "title": "EnquiryOut"
      },
      "EnquiryStatus": {
        "type": "string",
        "enum": [
          "new",
          "handled"
        ],
        "title": "EnquiryStatus"
      },
      "EnquiryUpdate": {
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "new",
              "handled"
            ],
            "title": "Status"
          }
        },
        "type": "object",
        "required": [
          "status"
        ],
        "title": "EnquiryUpdate"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "NewsletterIn": {
        "properties": {
          "email": {
            "type": "string",
            "maxLength": 320,
            "pattern": "^[^@\\s]+@[^@\\s]+\\.[^@\\s]+$",
            "title": "Email"
          },
          "source": {
            "type": "string",
            "maxLength": 40,
            "title": "Source",
            "default": "footer"
          }
        },
        "type": "object",
        "required": [
          "email"
        ],
        "title": "NewsletterIn"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "description": "Clerk session token",
        "scheme": "bearer"
      }
    }
  }
}